// src/Rest/model/tblMedico.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TblDepartamento } from './tblDepartamento';
import { TblPersona } from './tblPersona';
import { TblRol } from './tblRol';
import { Time } from './time';

export interface TblMedico { 
    idMedico?: number;
    medHoraEntrada?: Time;
    medHoraSalida?: Time;
    tblDepartamento?: TblDepartamento;
    tblPersona?: TblPersona;
    tblRol?: TblRol;
}